requirements.txt
pycuda
Pillow
argparse